elimenar material

+++ cliente/src/pages/Inventario/index.js
@@ -1,35 +1,53 @@
-import React, { useEffect, useState } from 'react'
+import React, { useCallback, useEffect, useMemo, useState } from 'react'
 import { Link } from 'react-router-dom'
 import TablaPaginacion from 'componentes/TablaPaginacion'
 import './index.css'
 import Material from 'services/materiales'
-// import { Form, Button } from 'react-bootstrap'
+import ModalConfirmacion from 'componentes/ModalConfirmacion'
 
 function Inventario(params) {
 
-    const [data,setData] = useState([])
+    const [eliminarConfirm, setEliminar] = useState(false)
+    const [showModalConfirmacion, setShowModalConfirmacion] = useState(false)
+    const [data, setData] = useState([])
+    const [id, setId] = useState(null)
+
     const m = new Material({})
 
-    useEffect(()=>{
+    useEffect(() => {
         m.getMateriales()
-            .then(m =>{
+            .then(m => {
                 setData(m)
             })
-    },[])
+    }, [])
 
-    const headers = ['#', 'Nombre', 'Descripcion', 'Existencias', 'Acciones']
+    const headers = useMemo(() => ['#', 'Nombre', 'Descripcion', 'Existencias', 'Acciones'], [])
 
-    const editar = ()=>{
+    const editar = useCallback(() => {
         alert('edicion')
+    }, [])
+
+    useEffect(() => {
+        if (eliminarConfirm) {
+            m._id = id
+            m.deleteMaterial(m)
+                .then(resp => {
+                    const newData = data.filter(s => s.IDMaterial !== id)
+                    setData(newData)
+                })
+        }
+    }, [eliminarConfirm])
+
+    const eliminar = useCallback((e) => {
+        handleCloseModalConfirmacion()
+        setId(parseInt(e.target.parentNode.id))
+        setEliminar(false)
+    }, [])
+
+    const handleCloseModalConfirmacion = () => {
+        setShowModalConfirmacion(!showModalConfirmacion)
     }
 
-    const eliminar = (e)=>{
-        console.log(parseInt(e.target.parentNode.id))
-        m._id = parseInt(e.target.parentNode.id)
-        m.deleteMaterial(m)
-            .then(resp=>console.log(resp))
-        // alert('Eliminar')
-    }
     return (
         <>
             <div className='container container-inventario'>
@@ -43,6 +61,13 @@ function Inventario(params) {
                         editar={editar}
                         eliminar={eliminar}
                         headers={headers}
+                    />
+                    <ModalConfirmacion
+                        show={showModalConfirmacion}
+                        body={"¿Esta seguro que desea elimiar el producto?"}
+                        title={"Eliminar Producto"}
+                        handleClose={handleCloseModalConfirmacion}
+                        respuesta={setEliminar}
                     />
                 </div>
             </div>
